refactor(test-runner): extract daft::styles into term-styles workspace crate
Moves the terminal ANSI styling utilities out of the daft crate so the
YAML test runner in xtask/ can use them without importing from daft. This
is the first step toward issue #516: drawing a clean seam between the
runner core and daft-specific bits so subsequent #509 sub-tasks
(clonefile, ramdisk, fixture cache, shared binary) can land as adapter
changes rather than accumulating coupling in the scheduler.

daft::styles becomes a thin re-export of term_styles so existing daft
callers compile unchanged. Runner imports switch to
`use term_styles as styles;`.
