+++ term-styles/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "term-styles"
+version = "0.1.0"
+edition = "2024"
+rust-version = "1.93"
+description = "Terminal ANSI styling utilities (colors, bold, dim) shared across the daft workspace."
+license = "MIT OR Apache-2.0"
+publish = false
+
+[lints.rust.unexpected_cfgs]
+level = "warn"
+check-cfg = ["cfg(daft_dev_build)"]
