+++ xtask/Cargo.toml
@@ -14,6 +14,7 @@ daft = { path = "..", version = "1.13.0" }
 rayon = "1.10"
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9"
+term-styles = { path = "../term-styles" }
 
 [target.'cfg(unix)'.dependencies]
 # Used by `xtask test-matrix` to detect parent-process death and self-terminate.
